requirements.txt
langchain
langchain-community
langchain-groq
sentence-transformers
docx2txt
pypdf
streamlit
python-dotenv